chat/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
node==1.2.2
oauthlib==3.2.2
odict==1.9.0
parso==0.8.4
pexpect==4.9.0
plumber==1.7
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.1.0
zope.lifecycleevent==5.0
zope.proxy==6.1
python-json-logger
python-decouple
python-logstash
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Pillow==9.0.1
psycopg2-binary

